fix class_utils dependency on pkg/types
